requirements.txt
langchain
langchain_community
langchain-openai
nbconvert
python-dotenv
numexpr
wikipedia